// src/interfaces/entities/sound.entity.ts
export interface ISound{
    _id: string;
    name: string;
    description: string;
    estimatedCost: {
      max: number;
      min: number;
    };
    list: boolean;
    createdAt: Date;
    updatedAt: Date;
}

export interface ISoundCreate {
    name: string;
    description: string;
    startingPrice: number;
    endingPrice: number;
    blocked?: boolean;
  }